move user test data to separate file
