+++ server/test/usersTestData.js
@@ -1,3 +1,59 @@
-const users = [];
+const users = [
+  {
+    name: 'Walter Okwa',
+    email: 'walter@gmail.com',
+    password: '1234567890',
+    confirmPassword: '1234567890',
+    role: 'caterer',
+  },
+  {
+    name: 'Ann Ihe',
+    email: 'annihe@gmail.com',
+    password: '1234567890',
+    confirmPassword: '1234567890',
+    role: 'customer',
+  },
+  {
+    name: 'Walter Okwa',
+    email: 'waltermenu@gmail.com',
+    password: '1234567890',
+    confirmPassword: '1234567890',
+    role: 'caterer',
+  },
+  {
+    name: 'Ann Ihe',
+    email: 'annihemenu@gmail.com',
+    password: '1234567890',
+    confirmPassword: '1234567890',
+    role: 'customer',
+  },
+  {
+    name: 'Walter Okwa',
+    email: 'walterorder@gmail.com',
+    password: '1234567890',
+    confirmPassword: '1234567890',
+    role: 'caterer',
+  },
+  {
+    name: 'Ann Ihe',
+    email: 'anniheorder@gmail.com',
+    password: '1234567890',
+    confirmPassword: '1234567890',
+    role: 'customer',
+  },
+  {
+    name: 'Solomon Okwa',
+    email: 'solozyokwa@gmail.com',
+    password: '1234567890',
+    confirmPassword: '1234567890',
+    role: 'caterer',
+  },
+  {
+    name: 'Walter Ifeanyi',
+    email: 'walterify@3d.com',
+    password: '1234567890',
+    role: 'customer',
+  },
+];
 
 export default users;
